requirements.txt
numpy
python-dotenv
mysql-connector-python
pymysql
scikit-learn
seaborn
catboost
xgboost
mlflow
pandas
SQLAlchemy
#-e .
